requirements.txt
constriction
numpy
pandas
pillow
pytorch-msssim
PyYAML
tqdm
vector-quantize-pytorch